% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APR_EAR.R
\name{APR_TO_EAR}
\alias{APR_TO_EAR}
\title{APR to EAR Calculator}
\usage{
APR_TO_EAR(.1, 12)
}
\arguments{
\item{APR}{the annual nominal rate of interest}

\item{compounding_period}{number of compounding periods in a year}
}
\value{
the effective return rate
}
\description{
Calculates the effective return rate from the annual nominal rate of interest and number of compounding periods
}
