% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_Ugly.R
\name{BIG_UGLY_PAYMENT}
\alias{BIG_UGLY_PAYMENT}
\title{Big Ugly for Present Value Calculator}
\usage{
BIG_UGLY_PAYMENT(19246.3, 30, .06, .03)

BIG_UGLY_PAYMENT(20401.08, 30, .06, .03, annuity = T)
}
\arguments{
\item{interest_rate}{rate of return for the compounding period}

\item{growth_rate}{inflation rate (reducing the value of the perpetuity)}

\item{annuity}{this indicates if the first payment is due immediately}

\item{payment}{st payment given at each compounding period}

\item{compounding_period}{number of compounding periods in a year}
}
\value{
the set payment
}
\description{
Calculates the set payment needed using a formula that Professor Gorman calls the Big Ugly.
}
