% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_Ugly.R
\name{BIG_UGLY_PV}
\alias{BIG_UGLY_PV}
\title{Big Ugly for Present Value Calculator}
\usage{
BIG_UGLY_PV(1000, 30, .06, .03)

BIG_UGLY_PV(1000, 30, .06, .03, annuity = T)
}
\arguments{
\item{payment}{st payment given at each compounding period}

\item{compounding_period}{number of compounding periods in a year}

\item{interest_rate}{rate or return for the compounding period}

\item{growth_rate}{inflation rate (reducing the value of the perpetuity)}

\item{annuity}{this indicates if the first payment is due immediately}
}
\value{
the present value
}
\description{
Calculates the present value using a formula that Professor Gorman calls the Big Ugly.
}
