% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APR_EAR.R
\name{EAR_to_APR}
\alias{EAR_to_APR}
\title{EAR to APR Calculator}
\usage{
APR_TO_EAR(.1, 12)
}
\arguments{
\item{EAR}{the effective return rate and number of compounding periods}

\item{compounding_period}{number of compounding periods in a year}
}
\value{
the annual nominal rate of interest
}
\description{
Calculates the annual nominal rate of interest from the effective return rate and number of compounding periods
}
