% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvfv.R
\name{FV}
\alias{FV}
\title{Future Value Calculator}
\usage{
FV(6.209213, 5, .1)
}
\arguments{
\item{periods}{amount of compounding intervals between the present to the future (typically in years)}

\item{interest_rate}{rate of return for the compounding period}

\item{fv}{Numerical value of a single good in the present (typically in $)}
}
\value{
the future Value of the single good
}
\description{
Calculates the Future Value of a given present value with an interest rate
}
