% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPV_IRR.R
\name{IRR}
\alias{IRR}
\title{IRR Calculator}
\usage{
NPV(c(0, 1, 2, 3, 0), .1)
}
\arguments{
\item{cashflows}{total incoming cash values in vector form for all compounding/payment periods (first element in vector is in present time)}

\item{interest_rate}{set rate of return for the compounding periods}
}
\value{
interest_rate for the compounding periods in order to be a "fair trade" (rounded)
}
\description{
Calculates the interest rate for the given cashflows to be a "fair trade"
}
