% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetuity.R
\name{PERPETIUTY_PV}
\alias{PERPETIUTY_PV}
\title{Perpetuity Present Value Calculator}
\usage{
PERPETIUTY_PV(1, .1, .08)
}
\arguments{
\item{payment}{Numerical value of the set payment forever}

\item{discount_rate}{rate of return for the compounding period}

\item{growth_rate}{inflation rate (reducing the value of the perpetuity)}
}
\value{
the present Value of the whole perpetuity
}
\description{
Calculates the Present Value of a given perpetuity with an interest and growth rate
}
