% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetuity.R
\name{PERPETIUTY_R}
\alias{PERPETIUTY_R}
\title{Perpetuity Interest Rate Calculator}
\usage{
PERPETIUTY_R(1, 50, .08)
}
\arguments{
\item{payment}{Numerical value of the set payment forever}

\item{PV}{the present value of the perpetuity}

\item{growth_rate}{inflation rate (reducing the value of the perpetuity)}
}
\value{
the discount rate of the whole perpetuity
}
\description{
Calculates the interest rate of a given perpetuity with the present value and growth rate
}
