% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvfv.R
\name{PV}
\alias{PV}
\title{Present Value Calculator}
\usage{
PV(10, 5, .1)
}
\arguments{
\item{fv}{Numerical value of a single good in a future time (typically in $)}

\item{periods}{amount of compounding intervals between the present to the future (typically in years)}

\item{interest_rate}{rate or return for the compounding period (typically in $)}
}
\value{
the present Value of the single good
}
\description{
Calculates the present Value of a given future value with an interest rate
}
